refactor: update relocated component references
