+++ website/.storybook/main.ts
@@ -1,7 +1,7 @@
 import type { StorybookConfig } from '@storybook/nextjs-vite';
 
 const config: StorybookConfig = {
-	stories: ['./*.mdx', '../src/stories/**/*.stories.@(ts|tsx)'],
+	stories: ['./*.mdx', '../src/**/*.stories.@(ts|tsx)'],
 	addons: ['@storybook/addon-docs', '@storybook/addon-a11y'],
 	framework: {
 		name: '@storybook/nextjs-vite',
