+++ website/src/app/[lang]/[region]/auth/confirm-login/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { useTranslator } from '@/lib/hooks/useTranslator';
 import { WebsiteLanguage } from '@/lib/i18n/utils';
 import { CONFIRM_LOGIN_PATH_REGEX } from '@/lib/utils/regex';
