+++ website/src/app/not-found.tsx
@@ -1,4 +1,4 @@
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { NotFound } from '@/components/not-found';
 import Link from 'next/link';
