+++ website/src/app/portal/delivery/overview/orange-money-csv-payout-process-dialog.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '@/components/dialog';
 import { StepResultBox } from '@/components/step-result-box';
 import {
