+++ website/src/app/portal/delivery/overview/payout-process-overview-client.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { DatePicker, normalizeToNoon } from '@/components/date-picker';
 import { getPayoutRecipientCountsAction } from '@/lib/server-actions/payout-process-actions';
 import { formatPayoutProcessLabel } from '@/lib/services/mobile-money-provider/payout-process-options';
