+++ website/src/app/portal/management/surveys/generate-surveys-dialog.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '@/components/dialog';
 import { StepResultBox } from '@/components/step-result-box';
 import { generateSurveysAction, previewSurveyGenerationAction } from '@/lib/server-actions/survey-actions';
