+++ website/src/app/portal/messaging/layout.tsx
@@ -1,5 +1,5 @@
 import { Breadcrumb } from '@/components/breadcrumb/breadcrumb';
-import { Card } from '@/components/card';
+import { Card } from '@/components/card/card';
 import { TabNavigation } from '@/components/tab-navigation';
 
 import { getAuthenticatedUserOrRedirect, requireAdmin } from '@/lib/firebase/current-user';
