+++ website/src/app/portal/messaging/templates/[sid]/send-message-wizard/recipients-table.tsx
@@ -1,8 +1,8 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { Checkbox } from '@/components/checkbox';
-import { Input } from '@/components/input';
+import { Input } from '@/components/input/input';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/popover';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/select';
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from '@/components/table';
