+++ website/src/app/portal/messaging/templates/[sid]/send-message-wizard/step-3-assignments.tsx
@@ -1,7 +1,7 @@
 'use client';
 
-import { Button } from '@/components/button';
-import { Input } from '@/components/input';
+import { Button } from '@/components/button/button';
+import { Input } from '@/components/input/input';
 import { RadioGroup, RadioGroupItem } from '@/components/radio-group';
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/select';
 import type { MessagingRecipientType } from '@/lib/services/twilio/messaging/recipients/recipients.types';
