+++ website/src/app/portal/messaging/templates/[sid]/send-message-wizard/wizard-footer.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { Loader2 } from 'lucide-react';
 import type { SendPhase } from './use-messaging-send';
