+++ website/src/app/portal/programs/[programId]/overview/components/donation-success-dialog.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { Button } from '@/components/button';
+import { Button } from '@/components/button/button';
 import { Dialog, DialogContent, DialogFooter, DialogHeader, DialogTitle } from '@/components/dialog';
 import { usePathname, useRouter, useSearchParams } from 'next/navigation';
