+++ website/src/app/portal/programs/[programId]/overview/components/stats-section.tsx
@@ -1,4 +1,4 @@
-import { Card } from '@/components/card';
+import { Card } from '@/components/card/card';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/components/tool-tip';
 import type { ProgramDashboardStats } from '@/lib/services/program-stats/program-stats.types';
 import { cn } from '@/lib/utils/cn';
