+++ website/src/app/portal/programs/[programId]/recipients/page.tsx
@@ -1,5 +1,5 @@
 import { BlockWrapper } from '@/components/block-wrapper';
-import { Card } from '@/components/card';
+import { Card } from '@/components/card/card';
 import { RecipientsTableClient } from '@/components/data-table/clients/recipients-table-client';
 import { tableQueryFromSearchParams } from '@/components/data-table/query-state';
 import { AppLoadingSkeleton } from '@/components/skeletons/app-loading-skeleton';
